requirements.txt
tensorflow
numpy
matplotlib
scikit-learn
pillow
streamlit
streamlit-drawable-canvas
streamlit-drawable-canvas